refactor the framework to a better structure.
1. application owns glib object, which takes charge of the
graphic-related tasks(directly interact with canvas and gl).

2. glib has a batches list, each element of which is a Batch object
(BasicBatch currently).Batch structure is used to manage the various gl states,
vao/vbos, indices, vertex attributes, shader uniforms, etc in a consecutive triangles
drawing process with gl.drawElements(). It has a Shader Object connected with it,
but it doesn't own that Shader Object.

2.1 Batch object defines the shape of a Rendering Functionality. by using
a concrete batch object, the question like how many vertex attributes will
be used, and what they are, and what data buffers in some certain formats
will be fed to them, are all declared. Also a bound Shader object must
conform to the requirements of the Connected Batch object.

3. Shader Object just manages the shader sources and a shader program.
It is responsible for shader compiling, attaching and linking. Although
a shader delcares the concrete various 'attributes', which must keep a
consitant relationship with the buffer data residing Batch object, the
Shader object itself has no ware of this. This awareness is reflected in
the Batch object outline.