add wireframe mode system in the engine.
Previously the batch and shader were fixed. they only contained positions
an indices data, and one BasicBatch can only bind a BasicShader with the
fixed shader files(vert.glsl and frag.glsl).

When it comes to drawing an object's wireframe', we use 'barycenter'
algorithm, in which way we need an extra data 'barycenter' for
interpolating from vertex shader to fragment shader.

In the same time we also need a mechanism to switch between 'wireframe'
mode and 'non-wireframe' mode smoothly. the mode can be specified in the
Scene Object.

So a little more extra work needs be done for refactoring the code.

The core of this work happens at "Primitive" part. when simple mesh
primitive is going to be created, we examine the rendering options, if
'wireframe' mode is set, then we modify the vertices data at this time
and rearrange them(vertices positions and indices) as well as adding a
new part 'barycenter'. all of the work is in 'wireframe.js'

Also the rendering option will be sent to glib's "sync" section, where
the shader's attributes will be updated('barycenter' attri is included)
if "wireframe" mode is found.